Rest: Reorganize code a bit
