+++ yamcs-core/src/main/java/org/yamcs/web/rest/InstanceRequestHandler.java
@@ -16,6 +16,8 @@ import org.yamcs.protobuf.YamcsManagement.ClientInfo.ClientState;
 import org.yamcs.protobuf.YamcsManagement.YamcsInstance;
 import org.yamcs.protobuf.YamcsManagement.YamcsInstances;
 import org.yamcs.web.rest.RestRequest.Option;
+import org.yamcs.web.rest.mdb.MDBRequestHandler;
+import org.yamcs.web.rest.processor.ProcessorRequestHandler;
 import org.yamcs.xtce.XtceDb;
 import org.yamcs.xtceproc.XtceDbFactory;
 
@@ -26,7 +28,7 @@ public class InstanceRequestHandler extends RestRequestHandler {
     final static Logger log = LoggerFactory.getLogger(InstanceRequestHandler.class.getName());
     
     @Override
-    protected RestResponse handleRequest(RestRequest req, int pathOffset) throws RestException {
+    public RestResponse handleRequest(RestRequest req, int pathOffset) throws RestException {
         if (!req.hasPathSegment(pathOffset)) {
             return listInstances(req);
         } else {
