+++ yamcs-core/src/main/java/org/yamcs/web/rest/RestUtils.java
@@ -10,13 +10,7 @@ import java.util.concurrent.TimeUnit;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.yamcs.protobuf.Yamcs.NamedObjectId;
 import org.yamcs.utils.TimeEncoding;
-import org.yamcs.xtce.Algorithm;
-import org.yamcs.xtce.MetaCommand;
-import org.yamcs.xtce.Parameter;
-import org.yamcs.xtce.SequenceContainer;
-import org.yamcs.xtce.XtceDb;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
@@ -78,6 +72,7 @@ public class RestUtils {
     public static ChannelFuture writeChunk(RestRequest req, ByteBuf buf) throws IOException {
         ChannelHandlerContext ctx = req.getChannelHandlerContext();
         Channel ch = ctx.channel();
+        log.debug("Writing a buf");
         if (!ch.isOpen()) {
             throw new IOException("Channel not or no longer open");
         }
@@ -128,193 +123,6 @@ public class RestUtils {
     
     public static IntervalResult scanForInterval(RestRequest req) throws RestException {
         return new IntervalResult(req);
-    }
-    
-    /**
-     * Searches for a valid parameter name in the URI of the request. It is
-     * assumed that the MDB for the instance was already added to the request's
-     * context.
-     * 
-     * @pathOffset the offset at which to start the search
-     */
-    public static MatchResult<Parameter> matchParameterName(RestRequest req, int pathOffset) {
-        XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        
-        MatchResult<String> nsMatch = matchXtceDbNamespace(req, pathOffset, false);
-        NamedObjectId id = null;
-        if (nsMatch.matches()) {
-            String namespace = nsMatch.getMatch();
-            if (req.hasPathSegment(nsMatch.getPathOffset())) {
-                String name = req.getPathSegment(nsMatch.getPathOffset());
-                id = NamedObjectId.newBuilder().setNamespace(namespace).setName(name).build();
-                Parameter p = findParameter(mdb, id);
-                if (p != null) {
-                    return new MatchResult<>(p, nsMatch.getPathOffset() + 1, id);
-                }
-            }
-        }
-
-        return new MatchResult<>(null, -1);
-    }
-    
-    public static Parameter findParameter(XtceDb mdb, NamedObjectId id) {
-        Parameter p = mdb.getParameter(id);
-        if(p==null) {
-            p = mdb.getSystemParameterDb().getSystemParameter(id, false);
-        }
-        return p;
-    }
-    
-    /**
-     * Searches for a valid container name in the URI of the request. It is
-     * assumed that the MDB for the instance was already added to the request's
-     * context.
-     * 
-     * @pathOffset the offset at which to start the search
-     */
-    public static MatchResult<SequenceContainer> matchContainerName(RestRequest req, int pathOffset) {
-        XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        
-        MatchResult<String> nsMatch = matchXtceDbNamespace(req, pathOffset, false);
-        NamedObjectId id = null;
-        if (nsMatch.matches()) {
-            String namespace = nsMatch.getMatch();
-            if (req.hasPathSegment(nsMatch.getPathOffset())) {
-                String name = req.getPathSegment(nsMatch.getPathOffset());
-                id = NamedObjectId.newBuilder().setNamespace(namespace).setName(name).build();
-                SequenceContainer c = mdb.getSequenceContainer(id);
-                if (c != null) {
-                    return new MatchResult<>(c, nsMatch.getPathOffset() + 1, id);
-                }
-            }
-        }
-
-        return new MatchResult<>(null, -1);
-    }
-    
-    /**
-     * Searches for a valid algorithm name in the URI of the request. It is
-     * assumed that the MDB for the instance was already added to the request's
-     * context.
-     * 
-     * @pathOffset the offset at which to start the search
-     */
-    public static MatchResult<Algorithm> matchAlgorithmName(RestRequest req, int pathOffset) {
-        XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        
-        MatchResult<String> nsMatch = matchXtceDbNamespace(req, pathOffset, false);
-        NamedObjectId id = null;
-        if (nsMatch.matches()) {
-            String namespace = nsMatch.getMatch();
-            if (req.hasPathSegment(nsMatch.getPathOffset())) {
-                String name = req.getPathSegment(nsMatch.getPathOffset());
-                id = NamedObjectId.newBuilder().setNamespace(namespace).setName(name).build();
-                Algorithm a = mdb.getAlgorithm(id);
-                if (a != null) {
-                    return new MatchResult<>(a, nsMatch.getPathOffset() + 1, id);
-                }
-            }
-        }
-
-        return new MatchResult<>(null, -1);
-    }
-    
-    /**
-     * Searches for a valid command name in the URI of the request. It is
-     * assumed that the MDB for the instance was already added to the request's
-     * context.
-     * 
-     * @pathOffset the offset at which to start the search
-     */
-    public static MatchResult<MetaCommand> matchCommandName(RestRequest req, int pathOffset) {
-        XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        
-        MatchResult<String> nsMatch = matchXtceDbNamespace(req, pathOffset, false);
-        NamedObjectId id = null;
-        if (nsMatch.matches()) {
-            String namespace = nsMatch.getMatch();
-            if (req.hasPathSegment(nsMatch.getPathOffset())) {
-                String name = req.getPathSegment(nsMatch.getPathOffset());
-                id = NamedObjectId.newBuilder().setNamespace(namespace).setName(name).build();
-                MetaCommand c = mdb.getMetaCommand(id);
-                if (c != null) {
-                    return new MatchResult<>(c, nsMatch.getPathOffset() + 1, id);
-                }
-            }
-        }
-
-        return new MatchResult<>(null, -1);
-    }
-    
-    /**
-     * Greedily matches a namespace
-     */
-    public static MatchResult<String> matchXtceDbNamespace(RestRequest req, int pathOffset, boolean strict) {
-        XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        String matchedNamespace = null;
-        
-        String segment = req.getPathSegment(pathOffset);
-        if (mdb.containsNamespace(segment)) {
-            matchedNamespace = segment;
-        } else if (mdb.containsNamespace("/" + segment)) {
-            matchedNamespace = "/" + segment; 
-        } else if (mdb.getSystemParameterDb().containsNamespace("/" + segment)) {
-            matchedNamespace = "/" + segment;
-        }
-        
-        int beyond = pathOffset;
-        if (matchedNamespace != null) {
-            beyond++;
-            if (matchedNamespace.startsWith("/")) {
-                for (int i = pathOffset+1; i < req.getPathSegmentCount(); i++) {
-                    String potential = matchedNamespace + "/" + req.getPathSegment(i);
-                    if (mdb.containsNamespace(potential) || mdb.getSystemParameterDb().containsNamespace(potential)) {
-                        matchedNamespace = potential;
-                        beyond++;
-                    }
-                }
-            }
-        }
-        
-        if (matchedNamespace != null && (!strict || beyond >= req.getPathSegmentCount())) {
-            return new MatchResult<>(matchedNamespace, beyond);
-        } else {
-            return new MatchResult<>(null, -1);
-        }
-    }
-    
-    public static class MatchResult <T> {
-        private final NamedObjectId requestedId;
-        private final T match;
-        private final int pathOffset; // positioned after the match
-        
-        MatchResult(T match, int pathOffset, NamedObjectId requestedId) {
-            this.match = match;
-            this.pathOffset = pathOffset;
-            this.requestedId = requestedId;
-        }
-        
-        MatchResult(T match, int pathOffset) {
-            this.match = match;
-            this.pathOffset = pathOffset;
-            requestedId = null;
-        }
-        
-        public boolean matches() {
-            return match != null;
-        }
-        
-        public NamedObjectId getRequestedId() {
-            return requestedId;
-        }
-        
-        public T getMatch() {
-            return match;
-        }
-        
-        public int getPathOffset() {
-            return pathOffset;
-        }
     }
     
     public static class IntervalResult {
