+++ yamcs-core/src/main/java/org/yamcs/web/rest/archive/ArchiveDownloadHandler.java
@@ -1,4 +1,4 @@
-package org.yamcs.web.rest;
+package org.yamcs.web.rest.archive;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -18,8 +18,24 @@ import org.yamcs.protobuf.Yamcs.NamedObjectId;
 import org.yamcs.protobuf.Yamcs.ReplayRequest;
 import org.yamcs.protobuf.Yamcs.TmPacketData;
 import org.yamcs.tctm.TmProviderAdapter;
+import org.yamcs.web.rest.BadRequestException;
+import org.yamcs.web.rest.NotFoundException;
+import org.yamcs.web.rest.ParameterReplayToChunkedCSVEncoder;
+import org.yamcs.web.rest.ParameterReplayToChunkedProtobufEncoder;
+import org.yamcs.web.rest.RestException;
+import org.yamcs.web.rest.RestParameterReplayListener;
+import org.yamcs.web.rest.RestRequest;
+import org.yamcs.web.rest.RestRequestHandler;
+import org.yamcs.web.rest.RestResponse;
+import org.yamcs.web.rest.RestStreams;
+import org.yamcs.web.rest.RestUtils;
 import org.yamcs.web.rest.RestUtils.IntervalResult;
-import org.yamcs.web.rest.RestUtils.MatchResult;
+import org.yamcs.web.rest.SqlBuilder;
+import org.yamcs.web.rest.StreamToChunkedCSVEncoder;
+import org.yamcs.web.rest.StreamToChunkedProtobufEncoder;
+import org.yamcs.web.rest.StreamToChunkedTransferEncoder;
+import org.yamcs.web.rest.mdb.MdbHelper;
+import org.yamcs.web.rest.mdb.MdbHelper.MatchResult;
 import org.yamcs.xtce.Parameter;
 import org.yamcs.yarch.TableDefinition;
 import org.yamcs.yarch.Tuple;
@@ -46,7 +62,7 @@ public class ArchiveDownloadHandler extends RestRequestHandler {
             switch (req.getPathSegment(pathOffset)) {
             case "parameters":
                 req.assertGET();
-                MatchResult<Parameter> mr = RestUtils.matchParameterName(req, pathOffset + 1);
+                MatchResult<Parameter> mr = MdbHelper.matchParameterName(req, pathOffset + 1);
                 if (!mr.matches()) {
                     throw new NotFoundException(req);
                 } else {
@@ -109,9 +125,8 @@ public class ArchiveDownloadHandler extends RestRequestHandler {
         }
         if (!nameSet.isEmpty()) {
             sqlb.where("pname in ('" + String.join("','", nameSet) + "')");
-        }        
+        }
         sqlb.descend(RestUtils.asksDescending(req, true));
-        
         String sql = sqlb.toString();
         
         if (req.asksFor(BINARY_MIME_TYPE)) {
@@ -143,21 +158,15 @@ public class ArchiveDownloadHandler extends RestRequestHandler {
             }
         }
         
-        StringBuilder sqlb = new StringBuilder("select ");
-        if (cols == null) {
-            sqlb.append("*");
-        } else if (cols.isEmpty()) {
-            throw new BadRequestException("No columns are specified.");
-        } else {
-            for (int i = 0; i < cols.size(); i++) {
-                if (i != 0) sqlb.append(", ");
-                sqlb.append(cols.get(i));
+        SqlBuilder sqlb = new SqlBuilder(table.getName());
+        if (cols != null) {
+            if (cols.isEmpty()) {
+                throw new BadRequestException("No columns were specified");
+            } else {
+                cols.forEach(col -> sqlb.select(col));
             }
         }
-        sqlb.append(" from ").append(table.getName());
-        if (RestUtils.asksDescending(req, false)) {
-            sqlb.append(" order desc");
-        }
+        sqlb.descend(RestUtils.asksDescending(req, false));
         String sql = sqlb.toString();
         
         RestStreams.stream(req, sql, new StreamToChunkedProtobufEncoder<TableRecord>(req, SchemaArchive.TableData.TableRecord.WRITE) {
@@ -179,24 +188,15 @@ public class ArchiveDownloadHandler extends RestRequestHandler {
             }
         }
 
-        StringBuilder sqlb = new StringBuilder("select * from ").append(EventRecorder.TABLE_NAME);
+        SqlBuilder sqlb = new SqlBuilder(EventRecorder.TABLE_NAME);
         IntervalResult ir = RestUtils.scanForInterval(req);
-        if (ir.hasInterval() || !sourceSet.isEmpty()) {
-            sqlb.append(" where ");
-            boolean first = true;
-            if (ir.hasInterval()) {
-                sqlb.append(ir.asSqlCondition("gentime"));
-                first = false;
-            }
-            if (!sourceSet.isEmpty()) {
-                if (!first) sqlb.append(" and ");
-                sqlb.append("source in ('").append(String.join("','", sourceSet)).append("')");
-                first = false;
-            }
+        if (ir.hasInterval()) {
+            sqlb.where(ir.asSqlCondition("gentime"));
         }
-        if (RestUtils.asksDescending(req, false)) {
-            sqlb.append(" order desc");
+        if (!sourceSet.isEmpty()) {
+            sqlb.where("source in ('" + String.join("','", sourceSet) + "')");
         }
+        sqlb.descend(RestUtils.asksDescending(req, false));
         String sql = sqlb.toString();
         
         if (req.asksFor(CSV_MIME_TYPE)) {
