+++ yamcs-core/src/main/java/org/yamcs/web/rest/mdb/MDBContainerRequestHandler.java
@@ -1,4 +1,4 @@
-package org.yamcs.web.rest;
+package org.yamcs.web.rest.mdb;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,8 +7,14 @@ import org.yamcs.protobuf.Rest.ListContainersResponse;
 import org.yamcs.protobuf.SchemaMdb;
 import org.yamcs.protobuf.SchemaRest;
 import org.yamcs.protobuf.Yamcs.NamedObjectId;
-import org.yamcs.web.rest.RestUtils.MatchResult;
+import org.yamcs.web.rest.NotFoundException;
+import org.yamcs.web.rest.RestException;
+import org.yamcs.web.rest.RestRequest;
+import org.yamcs.web.rest.RestRequestHandler;
+import org.yamcs.web.rest.RestResponse;
+import org.yamcs.web.rest.XtceToGpbAssembler;
 import org.yamcs.web.rest.XtceToGpbAssembler.DetailLevel;
+import org.yamcs.web.rest.mdb.MdbHelper.MatchResult;
 import org.yamcs.xtce.SequenceContainer;
 import org.yamcs.xtce.XtceDb;
 
@@ -24,7 +30,7 @@ public class MDBContainerRequestHandler extends RestRequestHandler {
         if (!req.hasPathSegment(pathOffset)) {
             return listContainers(req, null, mdb); // root namespace
         } else {
-            MatchResult<SequenceContainer> pm = RestUtils.matchContainerName(req, pathOffset);
+            MatchResult<SequenceContainer> pm = MdbHelper.matchContainerName(req, pathOffset);
             if (pm.matches()) { // container
                 return getSingleContainer(req, pm.getRequestedId(), pm.getMatch());
             } else { // namespace
@@ -35,7 +41,7 @@ public class MDBContainerRequestHandler extends RestRequestHandler {
     
     private RestResponse listContainersOrError(RestRequest req, int pathOffset) throws RestException {
         XtceDb mdb = req.getFromContext(MDBRequestHandler.CTX_MDB);
-        MatchResult<String> nsm = RestUtils.matchXtceDbNamespace(req, pathOffset, true);
+        MatchResult<String> nsm = MdbHelper.matchXtceDbNamespace(req, pathOffset, true);
         if (nsm.matches()) {
             return listContainers(req, nsm.getMatch(), mdb);
         } else {
