+++ yamcs-core/src/main/java/org/yamcs/web/rest/mdb/NameDescriptionSearchMatcher.java
@@ -1,4 +1,4 @@
-package org.yamcs.web.rest;
+package org.yamcs.web.rest.mdb;
 
 import java.util.Map.Entry;
